requirements.txt
pyyaml==6.0.2
mcp
httpx
openai
python-dotenv


